some Test are passing.Had to move the test folder to the project path.
